edited json seed data to include foreign keys
