+++ client/components/index.js
@@ -5,10 +5,8 @@
  */
 export {default as Navbar} from './navbar';
 export {default as UserHome} from './user-home';
-export {default as AllRestaurants} from './allRestaurants';
 export {default as SingleRestaurant} from './singleRestaurant';
 export {default as Filter} from './filter';
-export {default as MapView} from './map';
 export {Login, Signup} from './auth-form';
 export {default as HomePage} from './homePage';
 export {default as UserRestaurants} from './restaurantsVisited';
