+++ client/components/navbar.js
@@ -1,9 +1,5 @@
 import React, {Fragment} from 'react';
-import PropTypes from 'prop-types';
-import {connect} from 'react-redux';
-import {Link} from 'react-router-dom';
-import {logout} from '../store';
-import {Menu, Image, Icon, Responsive} from 'semantic-ui-react';
+import { Responsive} from 'semantic-ui-react';
 import styled from 'styled-components';
 import DesktopNavbar from './desktopNavBar';
 import MobileNavBar from './mobileNavBar';
