--- script/cuisines.json
@@ -1,1 +1,0 @@
-[{"id":1,"name":"chinese","createdAt":"2018-10-13T15:17:09.465Z","updatedAt":"2018-10-13T15:17:09.465Z"},{"id":4,"name":"mexican","createdAt":"2018-10-13T15:17:09.465Z","updatedAt":"2018-10-13T15:17:09.465Z"},{"id":6,"name":"burgers","createdAt":"2018-10-13T15:17:09.465Z","updatedAt":"2018-10-13T15:17:09.465Z"},{"id":7,"name":"steakhouse","createdAt":"2018-10-13T15:17:09.465Z","updatedAt":"2018-10-13T15:17:09.465Z"},{"id":2,"name":"italian","createdAt":"2018-10-13T15:17:09.465Z","updatedAt":"2018-10-13T15:17:09.465Z"},{"id":5,"name":"japanese","createdAt":"2018-10-13T15:17:09.465Z","updatedAt":"2018-10-13T15:17:09.465Z"},{"id":3,"name":"american","createdAt":"2018-10-13T15:17:09.464Z","updatedAt":"2018-10-13T15:17:09.464Z"}]
