--- server/db/models/cuisine.js
@@ -1,14 +1,0 @@
-const Sequelize = require('sequelize');
-const db = require('../db');
-
-const Cuisine = db.define('cuisines', {
-  name: {
-    type: Sequelize.STRING,
-    allowNull: false,
-    validate: {
-      notEmpty: true
-    }
-  }
-});
-
-module.exports = Cuisine;
