--- server/db/models/restaurant-cuisine.js
@@ -1,6 +1,0 @@
-const Sequelize = require('sequelize')
-const db = require('../db')
-
-const RestaurantCuisine = db.define('restaurant-cuisine', {})
-
-module.exports = RestaurantCuisine
